move compo function to internal
