+++ compo.go
@@ -2,7 +2,6 @@ package app
 
 import (
 	"net/url"
-	"strings"
 )
 
 // Compo is the interface that describes a component.
@@ -66,44 +65,10 @@ type CompoWithExtendedRender interface {
 	Funcs() map[string]interface{}
 }
 
-// ZeroCompo is the type to use as base for an empty compone.
+// ZeroCompo is the type to use as base for empty components.
 // Every instances of an empty struct is given the same memory address, which
 // causes problem for indexing components.
 // ZeroCompo have a placeholder field to avoid that.
 type ZeroCompo struct {
 	placeholder byte
-}
-
-// CompoNameFromURL is a helper function that returns the component name
-// targeted by the given URL.
-func CompoNameFromURL(u *url.URL) string {
-	if len(u.Scheme) != 0 && u.Scheme != "compo" {
-		return ""
-	}
-
-	path := u.Path
-	path = strings.TrimPrefix(path, "/")
-
-	paths := strings.SplitN(path, "/", 2)
-	if len(paths[0]) == 0 {
-		return ""
-	}
-	return normalizeCompoName(paths[0])
-}
-
-// CompoNameFromURLString is a helper function that returns the component
-// name targeted by the given URL.
-func CompoNameFromURLString(rawurl string) string {
-	u, _ := url.Parse(rawurl)
-	return CompoNameFromURL(u)
-}
-
-func normalizeCompoName(name string) string {
-	name = strings.ToLower(name)
-	if pkgsep := strings.IndexByte(name, '.'); pkgsep != -1 {
-		if name[:pkgsep] == "main" {
-			name = name[pkgsep+1:]
-		}
-	}
-	return name
 }
